requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap-v5==1.0.11
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
Pytess==1.0.0
pytesseract==0.3.10
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
